Refactor, improvements and other changes
