+++ components/ui/FilesPreview/index.vue
@@ -56,7 +56,7 @@ export default {
   display: grid;
   grid-template-columns: repeat(4, 1fr);
   grid-gap: 20px;
-  margin: 0 0 20px 0;
+
   &__item {
     transition: 0.5s;
     border-radius: 6px;
